You need to put the Substituin_Calculus_Pred in the bbin folder, also i dont think that will function.
Also, you need to change the directory of AtelierB and ProB in the files, callProB.py and solveroc.py. The bxml has to be on the folder, unless you change it in the main.py.

Update:
Now we get the predicate for implementations with operation calls. The operation call may have a deterministic machine, otherwise the implementation of the machine is searched, if not found it will raise an error. (We search the BXML files of the machine/implementation)
Added examples with operation calls